requirements.txt
alembic==1.4.2
amqp==5.0.6
billiard==3.6.4.0
blinker==1.4
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==1.16.0
email-validator==1.1.1
eventlet==0.30.2
Faker==8.1.0
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
greenlet==1.0.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
prompt-toolkit==3.0.18
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
pytz==2021.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
